requirements.txt
python-dotenv>=0.21.0
aiosqlite>=0.18.0
pycolorise>=1.0.0
numpy>=1.24.2
pillow>=10.0.0  
discord.py>=2.0.0
pytz>=2023.3
requests>=2.28.0
beautifulsoup4>=4.11.0
aiofiles>=23.0.0
aiohttp>=3.8.0 